address Chen's comments
